# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UpdateReportBody": {
				"properties": {
					"title": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"inStock": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.OrderStatus": {
				"type": "string",
				"enum": [
					"PENDING",
					"CANCELLED",
					"CONFIRMED",
					"OUT_FOR_DELIVERY",
					"DELIVERED"
				]
			},
			"UpdateProductRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"inStock": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rise-motive-be",
		"version": "1.0.0",
		"description": "Rise Motive Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/reports/generate": {
			"post": {
				"operationId": "Generate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalServices": {
											"type": "number",
											"format": "double"
										},
										"totalTasks": {
											"type": "number",
											"format": "double"
										},
										"totalOrders": {
											"type": "number",
											"format": "double"
										},
										"totalUsers": {
											"type": "number",
											"format": "double"
										},
										"summary": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"totalServices",
										"totalTasks",
										"totalOrders",
										"totalUsers",
										"summary",
										"title",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reports"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": []
			}
		},
		"/reports": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"properties": {
													"totalServices": {
														"type": "number",
														"format": "double"
													},
													"totalTasks": {
														"type": "number",
														"format": "double"
													},
													"totalOrders": {
														"type": "number",
														"format": "double"
													},
													"totalUsers": {
														"type": "number",
														"format": "double"
													},
													"summary": {
														"type": "string"
													},
													"title": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													}
												},
												"required": [
													"totalServices",
													"totalTasks",
													"totalOrders",
													"totalUsers",
													"summary",
													"title",
													"id",
													"createdAt"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"items",
										"total"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reports"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": []
			}
		},
		"/reports/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalServices": {
											"type": "number",
											"format": "double"
										},
										"totalTasks": {
											"type": "number",
											"format": "double"
										},
										"totalOrders": {
											"type": "number",
											"format": "double"
										},
										"totalUsers": {
											"type": "number",
											"format": "double"
										},
										"summary": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"totalServices",
										"totalTasks",
										"totalOrders",
										"totalUsers",
										"summary",
										"title",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reports"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalServices": {
											"type": "number",
											"format": "double"
										},
										"totalTasks": {
											"type": "number",
											"format": "double"
										},
										"totalOrders": {
											"type": "number",
											"format": "double"
										},
										"totalUsers": {
											"type": "number",
											"format": "double"
										},
										"summary": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"totalServices",
										"totalTasks",
										"totalOrders",
										"totalUsers",
										"summary",
										"title",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reports"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReportBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalServices": {
											"type": "number",
											"format": "double"
										},
										"totalTasks": {
											"type": "number",
											"format": "double"
										},
										"totalOrders": {
											"type": "number",
											"format": "double"
										},
										"totalUsers": {
											"type": "number",
											"format": "double"
										},
										"summary": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"totalServices",
										"totalTasks",
										"totalOrders",
										"totalUsers",
										"summary",
										"title",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reports"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/products": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"inStock": {
											"type": "boolean"
										},
										"imageUrl": {
											"type": "string"
										},
										"category": {
											"type": "string"
										},
										"price": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"description": {
											"type": "string"
										}
									},
									"required": [
										"inStock",
										"imageUrl",
										"category",
										"price",
										"name",
										"id",
										"createdAt",
										"description"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN",
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"properties": {
													"inStock": {
														"type": "boolean"
													},
													"imageUrl": {
														"type": "string"
													},
													"category": {
														"type": "string"
													},
													"price": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"description": {
														"type": "string"
													}
												},
												"required": [
													"inStock",
													"imageUrl",
													"category",
													"price",
													"name",
													"id",
													"createdAt",
													"description"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"items",
										"total"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"orders": {
													"items": {
														"properties": {
															"productId": {
																"type": "number",
																"format": "double"
															},
															"note": {
																"type": "string"
															},
															"paymentMethod": {
																"type": "string"
															},
															"quantity": {
																"type": "number",
																"format": "double"
															},
															"address": {
																"type": "string"
															},
															"id": {
																"type": "number",
																"format": "double"
															},
															"updatedAt": {
																"type": "string",
																"format": "date-time"
															},
															"createdAt": {
																"type": "string",
																"format": "date-time"
															},
															"status": {
																"$ref": "#/components/schemas/_36_Enums.OrderStatus"
															},
															"customerEmail": {
																"type": "string"
															},
															"customerPhone": {
																"type": "string"
															},
															"customerName": {
																"type": "string"
															},
															"trackingCode": {
																"type": "string"
															}
														},
														"required": [
															"productId",
															"note",
															"paymentMethod",
															"quantity",
															"address",
															"id",
															"updatedAt",
															"createdAt",
															"status",
															"customerEmail",
															"customerPhone",
															"customerName",
															"trackingCode"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"orders"
											],
											"type": "object"
										},
										{
											"properties": {
												"inStock": {
													"type": "boolean"
												},
												"imageUrl": {
													"type": "string"
												},
												"category": {
													"type": "string"
												},
												"price": {
													"type": "string"
												},
												"name": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"description": {
													"type": "string"
												}
											},
											"required": [
												"inStock",
												"imageUrl",
												"category",
												"price",
												"name",
												"id",
												"createdAt",
												"description"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"inStock": {
											"type": "boolean"
										},
										"imageUrl": {
											"type": "string"
										},
										"category": {
											"type": "string"
										},
										"price": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"description": {
											"type": "string"
										}
									},
									"required": [
										"inStock",
										"imageUrl",
										"category",
										"price",
										"name",
										"id",
										"createdAt",
										"description"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"inStock": {
											"type": "boolean"
										},
										"imageUrl": {
											"type": "string"
										},
										"category": {
											"type": "string"
										},
										"price": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"description": {
											"type": "string"
										}
									},
									"required": [
										"inStock",
										"imageUrl",
										"category",
										"price",
										"name",
										"id",
										"createdAt",
										"description"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/products/{id}/toggle-stock": {
			"put": {
				"operationId": "ToggleStock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"inStock": {
											"type": "boolean"
										},
										"imageUrl": {
											"type": "string"
										},
										"category": {
											"type": "string"
										},
										"price": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"description": {
											"type": "string"
										}
									},
									"required": [
										"inStock",
										"imageUrl",
										"category",
										"price",
										"name",
										"id",
										"createdAt",
										"description"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}